backend/requirements.txt
numpy
chromadb
openai
fastapi
fastembed
keybert
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
mysql-connector-python
pydantic
pymupdf
python-dotenv
requests
typing-extensions
python-multipart
uvicorn[standard]
